Update
Required-githooks: true


#Pragmas from previous commit message:
Quick-Functional: true
PR-repos: daos@PR-11447:15
RPM-test-version: 2.3.103
Skip-coverity-test: true
Skip-func-test-leap15: false
Skip-func-hw-test: true
Skip-scan-rpms: true
Skip-test-rpms: true
Test-tag: test_core_files
Skip-PR-comments: true
Skip-checkpatch: true
Skip-python-bandit: true
Skip-build: true
Quick-build: true
Quick-Functional: true
Skip-unit-tests: true
Skip-test: true
Allow-unstable-test: true
#RPM-test-version: version[-release]
#RPM-test-version: 2.3.103
# VM1-label: ci_vm1
# Ubuntu-VM9-label: ci_vm9
# Leap15-VM9-label: ci_vm9
# EL8-VM9-label: ci_vm9
Quick-Functional: true
Skip-coverity-test: true
Skip-func-test-leap15: false
Skip-func-hw-test: true
Skip-scan-rpms: true
Skip-test-rpms: true
Test-tag: test_core_files
Skip-PR-comments: true
